clean split
