+++ qiskit/quantum_info/__init__.py
@@ -115,8 +115,7 @@ Analysis
 
 from __future__ import annotations
 
-from qiskit._accelerate.sparse_observable import SparseObservable
-from qiskit._accelerate.py_sparse_observable import PySparseObservable
+from qiskit._accelerate.py_sparse_observable import SparseObservable
 
 from .analysis import hellinger_distance, hellinger_fidelity, Z2Symmetries
 from .operators import (
